extract inference_base, instantiate ort.InferenceSession independently
