+++ config/inference/pangu_rwrf_ckpt.yaml
@@ -1,3 +1,4 @@
+infer_type: "ckpt"
 best_ckpt: "./checkpoints/Pangu_240622_210919-epoch=225-val_loss_epoch=0.1788.ckpt"
 output_itv:
-  hours: 1
+  hours: 12
