+++ config/inference/pangu_rwrf_onnx.yaml
@@ -1,0 +1,5 @@
+infer_type: "onnx"
+onnx_path: "./export/Pangu_model.onnx"
+output_itv:
+  hours: 12
+gpu_id: 0
