+++ inference/__init__.py
@@ -1,1 +1,2 @@
-from .batch_inference import *
+from .batch_inference_ckpt import *
+from .batch_inference_onnx import *
