+++ visual/viz_radar.py
@@ -16,7 +16,7 @@ from .tw_background import TwBackground
 class VizRadar(TwBackground):
     def __init__(self):
         super().__init__()
-        self.title_suffix = "_radar_reflectivity"
+        self.title_prefix = "Radar Reflectivity\n"
 
     def plot_mxn(
         self,
@@ -46,6 +46,7 @@ class VizRadar(TwBackground):
 
         rows = 2  # gt/pred
         columns = ground_truth.shape[0]
+
         plt.close()
         fig, ax = plt.subplots(rows, columns, figsize=(20, 7), dpi=200, facecolor="w")
 
@@ -113,7 +114,7 @@ class VizRadar(TwBackground):
             zorder=0,
         )
         if title:
-            ax.set_title(title + self.title_suffix)
+            ax.set_title(self.title_prefix + title)
 
         # create an axes on the right side of ax. The width of cax will be 5%
         # of ax and the padding between cax and ax will be fixed at 0.05 inch.
@@ -134,20 +135,21 @@ class VizRadar(TwBackground):
         lon: np.ndarray,
         lat: np.ndarray,
         data: list[np.ndarray],
-        grid_on: bool = False,
         titles: list[str] = [],
+        grid_on: bool = False,
     ):
         if len(lat.shape) == 2 and len(lon.shape) == 2:
             lat = np.linspace(lat[0, 0], lat[-1, 0], lat.shape[0])
             lon = np.linspace(lon[0, 0], lon[0, -1], lon.shape[1])
 
         cols = len(data)
+
         plt.close()
         fig, ax = plt.subplots(1, cols, figsize=(20, 7), dpi=200, facecolor="w")
         for j in range(cols):
             tmp_ax = ax[j]
-            fig, tmp_ax = self.plot_bg(fig, tmp_ax, grid_on)
             title = titles[j] if titles else ""
+            fig, tmp_ax = self.plot_bg(fig, tmp_ax, grid_on)
             fig, tmp_ax = self._plot_radar(fig, tmp_ax, lon, lat, data[j], title)
 
         return fig, ax
