+++ visual/viz_vor.py
@@ -16,7 +16,7 @@ class VizVor(TwBackground):
     def __init__(self, pressure_level: int):
         super().__init__()
         self.press_lv = pressure_level
-        self.title_suffix = f"_vorticity@{self.press_lv}"
+        self.title_suffix = f"Vorticity@{self.press_lv}\n"
 
     def plot_mxn(
         self,
@@ -130,7 +130,7 @@ class VizVor(TwBackground):
             extend="both",
         )
         if title:
-            ax.set_title(title + self.title_suffix)
+            ax.set_title(self.title_suffix + title)
 
         # create an axes on the right side of ax. The width of cax will be 5%
         # of ax and the padding between cax and ax will be fixed at 0.05 inch.
@@ -141,6 +141,41 @@ class VizVor(TwBackground):
         cbar = fig.colorbar(conf, cax=cax)
         cbar.ax.set_title("($10^{-5} s^{-1}$)")
         cbar.set_ticks(np.arange(-100, 101, 20))
+
+        return fig, ax
+
+    def plot_1xn(
+        self,
+        lon: np.ndarray,
+        lat: np.ndarray,
+        u_wind_list: list[np.ndarray],
+        v_wind_list: list[np.ndarray],
+        grid_resolution_in_meter: list[int],
+        titles: list[str] = [],
+        grid_on: bool = False,
+    ):
+        if len(lat.shape) == 2 and len(lon.shape) == 2:
+            lat = np.linspace(lat[0, 0], lat[-1, 0], lat.shape[0])
+            lon = np.linspace(lon[0, 0], lon[0, -1], lon.shape[1])
+
+        cols = len(u_wind_list)
+
+        plt.close()
+        fig, ax = plt.subplots(1, cols, figsize=(20, 7), dpi=200, facecolor="w")
+        for j in range(cols):
+            tmp_ax = ax[j]
+            title = titles[j] if titles else ""
+            fig, tmp_ax = self.plot_bg(fig, tmp_ax, grid_on)
+            fig, tmp_ax = self._plot_vor(
+                fig,
+                tmp_ax,
+                lon,
+                lat,
+                u_wind_list[j],
+                v_wind_list[j],
+                grid_resolution_in_meter,
+                title,
+            )
 
         return fig, ax
